simp/submodules/face_generator/requirements.txt
opencv_python>=4.5.1.48
torch>=1.6.0
tqdm>=4.56.0
numpy>=1.19.5
matplotlib>=3.3.3
torchvision>=0.7.0
pandas>=1.1.5
Pillow>=8.2.0
PyYAML>=5.4.1
